feat: adding routeguard
