+++ frontend/src/pages/LandingPage.vue
@@ -1,7 +1,18 @@
 <script setup lang="ts">
-import Seeding from '../assets/images/seeding.png'
-import logo from '../assets/images/bc-gov-logo.png'
+import { useRouter } from 'vue-router'
 import login16 from '@carbon/icons-vue/es/login/16'
+import Seeding from '@/assets/images/seeding.png'
+import logo from '@/assets/images/bc-gov-logo.png'
+import AmplifyUserSession from '@/helpers/AmplifyUserSession'
+
+// extract the querystring parameters from the URL
+const router = useRouter()
+const { query } = router.currentRoute.value
+
+// check if a querystring parameter called ref exists and if it has a value of external
+if (query.ref && query.ref === 'external') {
+  AmplifyUserSession.logIn('bceidbusiness')
+}
 </script>
 
 <template>
@@ -14,9 +25,16 @@ import login16 from '@carbon/icons-vue/es/login/16'
           The Ministry of Forests' client management system
         </h2>
 
-        <p id="landing-desc" class="landing-desc">Check and manage client submissions</p>
+        <p id="landing-desc" class="landing-desc">
+          Check and manage client submissions
+        </p>
 
-        <bx-btn kind="primary" iconLayout="" class="landing-button">
+        <bx-btn
+          kind="primary"
+          iconLayout=""
+          class="landing-button"
+          @click.prevent="$session?.logIn('idir')"
+        >
           <span>Login with IDIR</span>
           <login16 slot="icon" />
         </bx-btn>
