+++ frontend/src/routes.ts
@@ -1,7 +1,5 @@
 import { createRouter, createWebHistory, type RouteRecordRaw } from 'vue-router'
 
-// The landing page
-import HomePage from '@/pages/HomePage.vue'
 // The Submission page, AKA internal user page
 import ReviewApplicationPage from '@/pages/ReviewApplicationPage.vue'
 // The external user page
@@ -10,18 +8,20 @@ import ApplyClientNumber from '@/pages/ApplyClientNumberPage.vue'
 import FormSubmittedPage from '@/pages/FormSubmittedPage.vue'
 // The page user will see while loading the login data
 import UserLoadingPage from '@/pages/UserLoadingPage.vue'
-// The page users will land when redirected to the application and not logged in
-import LoginPage from '@/pages/LoginPage.vue'
+// The landing page
 import LandingPage from '@/pages/LandingPage.vue'
+
+import AmplifyUserSession from '@/helpers/AmplifyUserSession'
 
 const routes: RouteRecordRaw[] = [
   {
-    path: '/landing', // Should be the root path in the future
+    path: '/landing',
     name: 'home',
     component: LandingPage,
     props: true,
     meta: {
-      hideHeader: true
+      hideHeader: true,
+      requireAuth: false
     }
   },
   {
@@ -29,37 +29,60 @@ const routes: RouteRecordRaw[] = [
     alias: '/',
     name: 'form',
     component: ApplyClientNumber,
-    props: true
+    props: true,
+    meta: {
+      hideHeader: false,
+      requireAuth: true
+    }
   },
   {
     path: '/form-submitted',
     name: 'confirmation',
     component: FormSubmittedPage,
-    props: true
+    props: true,
+    meta: {
+      hideHeader: false,
+      requireAuth: true
+    }
   },
   {
     path: '/submissions',
     name: 'internal',
     component: ReviewApplicationPage,
-    props: true
+    props: true,
+    meta: {
+      hideHeader: false,
+      requireAuth: true
+    }
   },
   {
     path: '/dashboard',
     name: 'loading',
     component: UserLoadingPage,
-    props: true
-  },
-  {
-    path: '/login',
-    name: 'login',
-    component: LoginPage,
-    props: true
+    props: true,
+    meta: {
+      hideHeader: true,
+      requireAuth: false
+    }
   }
 ]
 
 const router = createRouter({
   history: createWebHistory(),
   routes
+})
+
+/* This is a global guard that will run before each route change
+   It will check if the user is logged in and if the route requires auth
+  If the user is not logged in and the route requires auth, it will redirect to the home page for login
+*/
+router.beforeEach(async (to, from, next) => {
+  const user = await AmplifyUserSession.isLoggedIn()
+  if (to.meta.requireAuth && !user) {
+    next({ name: 'home' })
+  } else {
+    next()
+  }
 })
 
 export { routes, router }
